requirements.txt
# Core dependencies
numpy>=1.24.0
gymnasium>=0.29.0
minigrid>=2.3.0
pillow>=10.0.0
scipy>=1.11.0
matplotlib>=3.7.0

# For VLM integration (optional - only needed for real VLM experiments)
requests>=2.31.0

# For testing
pytest>=7.0.0
